Refactoring to split density map code, more javadocs
